different way of running examples as tests

+++ src/utils/initNewDay.ts
@@ -11,10 +11,16 @@ if (!day) {
 console.log(`Creating template for day ${day}`);
 
 const sourceFile = "templates/dayXX.ts";
+const sourceTestFile = "templates/dayXX.testtemplate.ts";
 const targetFile = `src/days/day${day.padStart(2, "0")}.ts`;
+const targetTestFile = `src/days/day${day.padStart(2, "0")}.test.ts`;
 
 if (!existsSync(sourceFile)) {
   console.error(`${sourceFile} does not exist, cannot init day ${day}`);
+  process.exit(1);
+}
+if (!existsSync(sourceTestFile)) {
+  console.error(`${sourceTestFile} does not exist, cannot init day ${day}`);
   process.exit(1);
 }
 
@@ -22,5 +28,10 @@ if (existsSync(targetFile)) {
   console.log(`day ${day} already exists`);
   process.exit(0);
 }
+if (existsSync(targetTestFile)) {
+  console.log(`day ${day} already exists`);
+  process.exit(0);
+}
 
 copyFileSync(sourceFile, targetFile);
+copyFileSync(sourceTestFile, targetTestFile);
